% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesDiagnostics-package.R
\docType{package}
\name{bayesDiagnostics-package}
\alias{bayesDiagnostics}
\alias{bayesDiagnostics-package}
\title{bayesDiagnostics: Bayesian Diagnostic Tools}
\description{
Comprehensive tools for diagnosing Bayesian models, including
posterior predictive checks, prior sensitivity analysis, and
MCMC diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ikrakib/bayesDiagnostics}
  \item Report bugs at \url{https://github.com/ikrakib/bayesDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Ibrahim Kholil Rakib \email{ikrakib1010@gmail.com}

}
\keyword{internal}
