% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesDiagnostics-package.R
\docType{package}
\name{bayesDiagnostics}
\alias{bayesDiagnostics-package}
\alias{bayesDiagnostics}
\title{bayesDiagnostics Package}
\description{
Provides comprehensive tools for Bayesian model diagnostics and comparison.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ikrakib/bayesDiagnostics}
  \item Report bugs at \url{https://github.com/ikrakib/bayesDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Ibrahim Kholil Rakib \email{ikrakib1010@gmail.com}

}
