% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_check.R
\name{calculate_bayesian_pvalues}
\alias{calculate_bayesian_pvalues}
\title{Calculate Bayesian P-values}
\usage{
calculate_bayesian_pvalues(observed_stats, replicated_stats)
}
\description{
Calculate Bayesian P-values
}
\keyword{internal}
