% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size_diagnostics.R
\name{calculate_per_chain_ess}
\alias{calculate_per_chain_ess}
\title{Calculate Per-Chain ESS}
\usage{
calculate_per_chain_ess(model, parameters)
}
\description{
Calculate Per-Chain ESS
}
\keyword{internal}
