% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_sensitivity.R
\name{calculate_sensitivity_metrics}
\alias{calculate_sensitivity_metrics}
\title{Calculate Sensitivity Metrics}
\usage{
calculate_sensitivity_metrics(original, refitted_list, parameters, metric)
}
\description{
Calculate Sensitivity Metrics
}
