% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_check.R
\name{calculate_test_statistics}
\alias{calculate_test_statistics}
\title{Calculate Test Statistics}
\usage{
calculate_test_statistics(data, test_statistics)
}
\description{
Calculate Test Statistics
}
\keyword{internal}
