% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_sensitivity.R
\name{calculate_wasserstein_distance}
\alias{calculate_wasserstein_distance}
\title{Wasserstein Distance Calculation}
\usage{
calculate_wasserstein_distance(x, y)
}
\description{
Wasserstein Distance Calculation
}
\keyword{internal}
