% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{check_parameter_convergence}
\alias{check_parameter_convergence}
\title{Check Parameter Convergence}
\usage{
check_parameter_convergence(params, rhat_threshold, ess_threshold)
}
\description{
Check Parameter Convergence
}
\keyword{internal}
