% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{collect_convergence_warnings}
\alias{collect_convergence_warnings}
\title{Collect Convergence Warnings}
\usage{
collect_convergence_warnings(
  pop_diag,
  group_diag,
  shrinkage,
  rhat_threshold,
  ess_threshold
)
}
\description{
Collect Convergence Warnings
}
\keyword{internal}
