% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_robustness.R
\name{compute_robustness_index}
\alias{compute_robustness_index}
\title{Compute robustness index}
\usage{
compute_robustness_index(sensitivity_surfaces)
}
\description{
Compute robustness index
}
\keyword{internal}
