% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_robustness.R
\name{compute_sensitivity_metric}
\alias{compute_sensitivity_metric}
\title{Compute sensitivity metric}
\usage{
compute_sensitivity_metric(original, perturbed, metric_type)
}
\description{
Compute sensitivity metric
}
\keyword{internal}
