% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{create_convergence_plots}
\alias{create_convergence_plots}
\title{Create Convergence Plots}
\usage{
create_convergence_plots(result, ...)
}
\description{
Create Convergence Plots
}
\keyword{internal}
