% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{create_elicitation_diagnostics}
\alias{create_elicitation_diagnostics}
\title{Create elicitation diagnostics}
\usage{
create_elicitation_diagnostics(result, data_sample)
}
\description{
Create elicitation diagnostics
}
\keyword{internal}
