% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{create_sensitivity_guidance}
\alias{create_sensitivity_guidance}
\title{Create sensitivity guidance}
\usage{
create_sensitivity_guidance(recommended, alternatives)
}
\description{
Create sensitivity guidance
}
\keyword{internal}
