% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{elicit_continuous_prior}
\alias{elicit_continuous_prior}
\title{Elicit continuous prior}
\usage{
elicit_continuous_prior(plausible_range, most_likely, confidence)
}
\description{
Elicit continuous prior
}
\keyword{internal}
