% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{elicit_discrete_prior}
\alias{elicit_discrete_prior}
\title{Elicit discrete prior}
\usage{
elicit_discrete_prior(plausible_range, most_likely)
}
\description{
Elicit discrete prior
}
\keyword{internal}
