% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{elicit_proportion_prior}
\alias{elicit_proportion_prior}
\title{Elicit proportion prior}
\usage{
elicit_proportion_prior(most_likely, confidence)
}
\description{
Elicit proportion prior
}
\keyword{internal}
