% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size_diagnostics.R
\name{extract_all_parameters}
\alias{extract_all_parameters}
\title{Extract All Parameters}
\usage{
extract_all_parameters(model)
}
\description{
Extract All Parameters
}
\keyword{internal}
