% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size_diagnostics.R
\name{extract_ess_statistics}
\alias{extract_ess_statistics}
\title{Extract ESS Statistics}
\usage{
extract_ess_statistics(model, parameters)
}
\description{
Extract ESS Statistics
}
\keyword{internal}
