% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_unified.R
\name{extract_from_brms}
\alias{extract_from_brms}
\title{Extract from brmsfit}
\usage{
extract_from_brms(model, parameters, include_warmup, chains)
}
\description{
Extract from brmsfit
}
\keyword{internal}
