% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_unified.R
\name{extract_from_cmdstan}
\alias{extract_from_cmdstan}
\title{Extract from CmdStanMCMC}
\usage{
extract_from_cmdstan(model, parameters, include_warmup, chains)
}
\description{
Extract from CmdStanMCMC
}
\keyword{internal}
