% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_unified.R
\name{extract_from_mcmc_list}
\alias{extract_from_mcmc_list}
\title{Extract from mcmc.list}
\usage{
extract_from_mcmc_list(model, parameters, chains)
}
\description{
Extract from mcmc.list
}
\keyword{internal}
