% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_unified.R
\name{extract_from_stanfit}
\alias{extract_from_stanfit}
\title{Extract from stanfit}
\usage{
extract_from_stanfit(model, parameters, include_warmup, chains)
}
\description{
Extract from stanfit
}
\keyword{internal}
