% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_unified.R
\name{extract_from_stanreg}
\alias{extract_from_stanreg}
\title{Extract from stanreg}
\usage{
extract_from_stanreg(model, parameters, include_warmup, chains)
}
\description{
Extract from stanreg
}
\keyword{internal}
