% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_robustness.R
\name{extract_posterior_summary}
\alias{extract_posterior_summary}
\title{Extract posterior summary}
\usage{
extract_posterior_summary(model, parameters, credible_level)
}
\description{
Extract posterior summary
}
\keyword{internal}
