% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size_diagnostics.R
\name{generate_ess_recommendations}
\alias{generate_ess_recommendations}
\title{Generate ESS Recommendations}
\usage{
generate_ess_recommendations(problematic, ess_stats)
}
\description{
Generate ESS Recommendations
}
\keyword{internal}
