% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_check.R
\name{generate_posterior_predictive_samples}
\alias{generate_posterior_predictive_samples}
\title{Generate Posterior Predictive Samples}
\usage{
generate_posterior_predictive_samples(model, n_samples)
}
\description{
Generate Posterior Predictive Samples
}
