% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{generate_prior_alternatives}
\alias{generate_prior_alternatives}
\title{Generate prior alternatives}
\usage{
generate_prior_alternatives(recommended, param_type, confidence)
}
\description{
Generate prior alternatives
}
\keyword{internal}
