% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_robustness.R
\name{generate_prior_perturbations}
\alias{generate_prior_perturbations}
\title{Generate prior perturbations}
\usage{
generate_prior_perturbations(priors, direction, dimensions)
}
\description{
Generate prior perturbations
}
\keyword{internal}
