% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_robustness.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations}
\usage{
generate_recommendations(concerning_parameters, priors)
}
\description{
Generate recommendations
}
\keyword{internal}
