% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{generate_stronger_version}
\alias{generate_stronger_version}
\title{Generate stronger version}
\usage{
generate_stronger_version(prior_obj, param_type)
}
\description{
Generate stronger version
}
\keyword{internal}
