% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation_helper.R
\name{generate_weaker_version}
\alias{generate_weaker_version}
\title{Generate weaker version}
\usage{
generate_weaker_version(prior_obj, param_type)
}
\description{
Generate weaker version
}
\keyword{internal}
