% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{identify_group_params}
\alias{identify_group_params}
\title{Identify Group Parameters}
\usage{
identify_group_params(diagnostics, group_vars)
}
\description{
Identify Group Parameters
}
\keyword{internal}
