% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{identify_population_params}
\alias{identify_population_params}
\title{Identify Population Parameters}
\usage{
identify_population_params(diagnostics)
}
\description{
Identify Population Parameters
}
\keyword{internal}
