% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_comparison.R
\name{interpret_bayes_factor}
\alias{interpret_bayes_factor}
\title{Interpret Bayes Factor Strength}
\usage{
interpret_bayes_factor(bf)
}
\description{
Interpret Bayes Factor Strength
}
\keyword{internal}
