% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_comparison.R
\name{plot.bayes_factor_comparison}
\alias{plot.bayes_factor_comparison}
\title{Plot Bayes Factor Comparison Results}
\usage{
\method{plot}{bayes_factor_comparison}(x, ...)
}
\arguments{
\item{x}{A bayes_factor_comparison object}

\item{...}{Additional arguments passed to ggplot2 functions}
}
\value{
A ggplot2 plot object
}
\description{
Plot Bayes Factor Comparison Results
}
