% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_convergence.R
\name{summarize_convergence}
\alias{summarize_convergence}
\title{Summarize Convergence}
\usage{
summarize_convergence(pop_diag, group_diag, shrinkage)
}
\description{
Summarize Convergence
}
\keyword{internal}
